change to display image edit/delete error message for non-owners

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -26,6 +26,8 @@ public class ImageController {
 
     @Autowired
     private TagService tagService;
+    public  static final String EDIT_IMAGE_ERROR ="Only the owner of the image can edit the image";
+    public  static final String DELETE_IMAGE_ERROR ="Only the owner of the image can delete the image";
 
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
@@ -92,13 +94,19 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        User currentUser= (User) session.getAttribute("loggeduser");
+        if(currentUser.getId()==image.getUser().getId()){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else {
+            model.addAttribute("image",image);
+            model.addAttribute("editError", EDIT_IMAGE_ERROR);
+            return  "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +148,18 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session, Model model) {
+        Image image=imageService.getImage(imageId);
+        User currentUser= (User) session.getAttribute("loggeduser");
+        if(currentUser.getId()==image.getUser().getId()){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else{
+            model.addAttribute("image",image);
+            model.addAttribute("deleteError", DELETE_IMAGE_ERROR);
+            return  "images/image";
+        }
+
     }
